ansible/roles/codebase_dr/files/DisasterRecord/requirements.txt
absl-py==0.1.13
affine==2.2.0
astor==0.6.2
attrs==17.4.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
boto3==1.5.32
botocore==1.8.46
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
docutils==0.14
elasticsearch==6.3.1
Flask==1.0.2
#functools32==3.2.3.post2
#futures==3.2.0
geographiclib==1.49
Geohash==1.0
geopy==1.17.0
html5lib==0.9999999
ibm-cos-sdk==2.3.0
ibm-cos-sdk-core==2.3.0
ibm-cos-sdk-s3transfer==2.3.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mkl-fft==1.0.6
#mkl-random==1.0.1
mkl-random
networkx==2.2
nltk==3.3
numpy==1.15.1
Pillow==5.2.0
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
requests==2.18.4
s3transfer==0.1.13
scikit-image==0.14.0
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.2
toolz==0.9.0
tornado==5.1
urllib3==1.23
Werkzeug==0.14.1
wordcloud==1.5.0